--- a/changelog/@unreleased/pr-20.v2.yml
+++ b/changelog/@unreleased/pr-20.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Microbenchmark Path creation
+  links:
+  - https://github.com/palantir/syntactic-paths/pull/20