--- a/syntactic-paths/src/jmh/java/com/palantir/util/syntacticpath/PathsBenchmark.java
+++ b/syntactic-paths/src/jmh/java/com/palantir/util/syntacticpath/PathsBenchmark.java
@@ -1,0 +1,65 @@
+/*
+ * (c) Copyright 2022 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.util.syntacticpath;
+
+import java.util.concurrent.TimeUnit;
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.BenchmarkMode;
+import org.openjdk.jmh.annotations.Fork;
+import org.openjdk.jmh.annotations.Measurement;
+import org.openjdk.jmh.annotations.Mode;
+import org.openjdk.jmh.annotations.OutputTimeUnit;
+import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.State;
+import org.openjdk.jmh.annotations.Threads;
+import org.openjdk.jmh.annotations.Warmup;
+
+@BenchmarkMode(Mode.AverageTime)
+@OutputTimeUnit(TimeUnit.MICROSECONDS)
+@Warmup(iterations = 2, time = 1, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 3, time = 1, timeUnit = TimeUnit.SECONDS)
+@Fork(1)
+@Threads(Threads.MAX)
+@State(Scope.Benchmark)
+@SuppressWarnings({"designforextension", "NullAway"})
+public class PathsBenchmark {
+
+    @Benchmark
+    public Path stringOne() {
+        return Paths.get("/test");
+    }
+
+    @Benchmark
+    public Path stringThree() {
+        return Paths.get("/test/foo/bar");
+    }
+
+    @Benchmark
+    public Path varArgsOne() {
+        return Paths.get("/test", "");
+    }
+
+    @Benchmark
+    public Path varArgsTwo() {
+        return Paths.get("/test", "foo");
+    }
+
+    @Benchmark
+    public Path varArgsThree() {
+        return Paths.get("/test", "foo", "bar");
+    }
+}