--- a/versions.lock
+++ b/versions.lock
@@ -1,5 +1,5 @@
 # Run ./gradlew --write-locks to regenerate this file
-com.fasterxml.jackson.core:jackson-annotations:2.13.3 (1 constraints: 89123821)
+com.fasterxml.jackson.core:jackson-annotations:2.13.3 (2 constraints: c3178471)
 com.google.code.findbugs:jsr305:3.0.2 (1 constraints: 170aecb4)
 com.google.errorprone:error_prone_annotations:2.11.0 (2 constraints: 751be887)
 com.google.guava:failureaccess:1.0.1 (1 constraints: 140ae1b4)