--- a/versions.props
+++ b/versions.props
@@ -1,7 +1,8 @@
-com.fasterxml.jackson.core:jackson-databind = 2.13.3
+com.fasterxml.jackson.*:* = 2.13.3
 com.google.guava:guava = 31.1-jre
 com.palantir.safe-logging:* = 3.0.0
 org.assertj:assertj-core = 3.23.1
 org.junit.jupiter:* = 5.9.0
 org.mockito:* = 4.8.0
+org.openjdk.jmh:* = 1.35
 org.slf4j:* = 1.7.36