## Before this PR
There were no microbenchmarks for various `Path` creation methods.
```
# VM version: JDK 11.0.16.1, OpenJDK 64-Bit Server VM, 11.0.16.1+9-LTS
Benchmark                    Mode  Cnt  Score   Error  Units
PathsBenchmark.stringOne     avgt    3  0.078 ± 0.067  us/op
PathsBenchmark.stringThree   avgt    3  0.135 ± 0.047  us/op
PathsBenchmark.varArgsOne    avgt    3  0.108 ± 0.117  us/op
PathsBenchmark.varArgsThree  avgt    3  0.242 ± 0.354  us/op
PathsBenchmark.varArgsTwo    avgt    3  0.216 ± 0.472  us/op
```

```
# VM version: JDK 17.0.4.1, OpenJDK 64-Bit Server VM, 17.0.4.1+9-LTS
Benchmark                    Mode  Cnt  Score   Error  Units
PathsBenchmark.stringOne     avgt    3  0.063 ± 0.019  us/op
PathsBenchmark.stringThree   avgt    3  0.128 ± 0.037  us/op
PathsBenchmark.varArgsOne    avgt    3  0.089 ± 0.026  us/op
PathsBenchmark.varArgsThree  avgt    3  0.217 ± 0.075  us/op
PathsBenchmark.varArgsTwo    avgt    3  0.198 ± 0.021  us/op
```


## After this PR
```
# VM version: JDK 11.0.16.1, OpenJDK 64-Bit Server VM, 11.0.16.1+9-LTS
Benchmark                    Mode  Cnt  Score   Error  Units
PathsBenchmark.stringOne     avgt    3  0.063 ± 0.115  us/op
PathsBenchmark.stringThree   avgt    3  0.116 ± 0.067  us/op
PathsBenchmark.varArgsOne    avgt    3  0.094 ± 0.018  us/op
PathsBenchmark.varArgsThree  avgt    3  0.199 ± 0.027  us/op
PathsBenchmark.varArgsTwo    avgt    3  0.148 ± 0.050  us/op
```

```
# VM version: JDK 17.0.4.1, OpenJDK 64-Bit Server VM, 17.0.4.1+9-LTS
Benchmark                    Mode  Cnt  Score   Error  Units
PathsBenchmark.stringOne     avgt    3  0.061 ± 0.023  us/op
PathsBenchmark.stringThree   avgt    3  0.111 ± 0.047  us/op
PathsBenchmark.varArgsOne    avgt    3  0.094 ± 0.010  us/op
PathsBenchmark.varArgsThree  avgt    3  0.210 ± 0.119  us/op
PathsBenchmark.varArgsTwo    avgt    3  0.155 ± 0.049  us/op
```
==COMMIT_MSG==
Microbenchmark Path creation

==COMMIT_MSG==

## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->